gateway/pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "circuitbreaker>=2.1.3",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.8.1",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "uvicorn>=0.34.1",
]
